update error message screen
